requirements.txt

pandas>=1.5.0
numpy>=1.20.0

dvc>=2.0.0
PyYAML>=6.0

scikit-learn>=1.1.0
joblib>=1.1.0

scipy>=1.8.0        # Para matrices dispersas (csr_matrix, save_npz)

mlxtend>=0.18.0

matplotlib>=3.5.0   # Para plots en evaluate.py
shap>=0.40.0        # Para SHAP values en evaluate.py

python-dateutil>=2.8.0 # Para relativedelta en train_recommender.py

pyarrow>=10.0.0

jupyterlab
# notebook